CSC 113: Computer Programming Semester Project
The program is written in C++ with a command-line interface (CLI). The
design and ensuing program are modular in nature and make maximum use
of abstract data types and of software re-use. Particular attention is
paid to details of the program from coding to the final product.

In order to run the coffeehouse effectively, cost and time saving is
essential. The program Dunkin’ Donuts Easy Order shall do just about
that. It is a computerized easy order system for customers to quickly
give orders through a system to the restaurant and know the total
amount instantly and effectively.